pyproject.toml
[tool.poetry]
name = "mailing_manager"
version = "0.1.0"
description = "Basic mail template managing integrated with admin panel and django-mail-queue."
authors = ["Codi Cooperatiu <hola@codi.coop>"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.6"
django-mail-queue = "^3.2.4"
Django = "^3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
